refactor(workspace): type FITS write dimensions as the lengths they describe
The `rp_fits` writer took `width`/`height` as `u32` because they end up
as `NAXIS1`/`NAXIS2` header cards. But they are also the length the
pixel buffer is validated against, so the function was converting on
every call, and `FitsError::DimensionMismatch` carried `got: usize` and
`expected: usize` beside `width: u32`.

Taking `usize` puts the conversion at the last consumer that needs a
fixed width instead of the first function that touches one. The capture
path stops narrowing `image_array.dim()` to `u32` only to widen it
straight back for the `Array2` shape; the one narrowing left is the JSON
sidecar, which is genuinely serialized. The `i64::try_from` this adds on
the `NAXIS` side is offset by the `checked_mul` overflow arm becoming
reachable — dead code on every 64-bit target with `u32` parameters, and
now a two-line test.

`usize` parameters do mean a caller can ask for a product no buffer
could hold, so the wrapper checks in `rp::persistence` and
`phd2_guider` move from `*` to `checked_mul`; each has a test.

Two boundaries stay where they are. `sky_survey_camera_harness` holds
dimensions that are both a `vec![0u16; n]` length and `f64` WCS header
cards, and `f64::from(usize)` does not exist, so it converts
explicitly. `gen_autofocus_fixtures` keeps `u32` consts because its
star loops work in signed offsets.

Production `x as usize` in the capture/write path: 12 -> 3 (one doc
comment, two in the example's signed index arithmetic, which belongs
with the float and sign-change buckets).

+++ crates/rp-fits/src/error.rs
@@ -13,8 +13,8 @@ pub enum FitsError {
     #[error("dimension mismatch: pixel count {got} does not match {width}x{height} (expected {expected})")]
     DimensionMismatch {
         got: usize,
-        width: u32,
-        height: u32,
+        width: usize,
+        height: usize,
         expected: usize,
     },
